Edited

+++ CASPortal/Models/TimeSlot.cs
@@ -9,9 +9,9 @@ namespace CASPortal.Models
     {
         public int ItemID { get; set; }
         public string Date { get; set; }
-        public int BusinessStartHour { get; set; }
-        public int BusinessEndHour { get; set; }
-        public int StartTime { get; set; }
-        public int EndTime { get; set; }
+        public float BusinessStartHour { get; set; }
+        public float BusinessEndHour { get; set; }
+        public float StartTime { get; set; }
+        public float EndTime { get; set; }
     }
 }
